alter in Somar_N_m,

+++ atvd6/SOMAR_N_M.C
@@ -11,7 +11,7 @@ int Sum(int value){
     }
 }
 
-double fat(int n)
+double fat(double n)
 {
   double vfat;
   if ( n <= 1 )
